dashboard/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
black==19.3b0
cffi==1.12.2
Click==7.0
cryptography==2.6.1
Django==2.1.7
django-rest-knox==3.6.0
djangorestframework==3.9.2
mysqlclient==1.4.2
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2018.9
six==1.12.0
toml==0.10.0
yapf==0.26.0
